broke all database connections, checking previous commit for hints

+++ util/clean-dev.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
 
 #### cleanup docker images, best to run this once a week or so
-docker rmi -f $(docker images -qf dangling=true)
+DANGLING_IMGS=$(docker images -qf dangling=true)
+
+if [ ! "$DANGLING_IMGS" ]; then 
+    echo "Dev is clean"
+else
+    docker rmi -f $(docker images -qf dangling=true)
+    echo "Dev is clean"
+fi
